+++ lib/src/helpers/time.dart
@@ -6,7 +6,9 @@ class IRTimeModel {
   int hour;
   int minute;
 
-  IRTimeModel({required this.hour, required this.minute});
+  IRTimeModel({required this.hour, required this.minute})
+      : assert(0 <= hour && hour <= 23, "hour is not valid"),
+        assert(0 <= minute && minute <= 59, "minute is not valid");
 
   @override
   String toString() {